% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstImp.R
\name{firstImp}
\alias{firstImp}
\title{Imputation}
\usage{
firstImp(sepdata, maxiter = 1)
}
\arguments{
\item{sepdata}{A data frame with missing values.}

\item{maxiter}{Max number of iterations for missForest imputation.}
}
\value{
Returns imputed data.
}
\description{
Imputation using missForest package for the big data. This is to prepare for case distance calculation or impute the artificial missingness.
}
