% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctMatch.R
\name{pctMatch}
\alias{pctMatch}
\title{Percent Identical Feature Values}
\usage{
pctMatch(originaldata, sifteddata)
}
\arguments{
\item{originaldata}{Original data frame.}

\item{sifteddata}{Output data frame from DataSifter.}
}
\value{
Returns a list of percentages for all rows.
}
\description{
Measures the Percent of Identical Feature Values (PIFV) between the “sifted” record and the original record.
}
