% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjdist.R
\name{subjdist}
\alias{subjdist}
\title{Subject Distance}
\usage{
subjdist(completedata)
}
\arguments{
\item{completedata}{A complete data frame without missing values.}
}
\value{
Return a dataframe that shows sorted between subject distances according to structured variables.
}
\description{
Calculates the Euclidian distance between two subjects based on their feature values.
}
